fix(session): set secure and sameSite to session cookies
Prevent cookie from being accessed over http in staging or production, and set sameSite to strict
(default lax) to mitigate any potential CSRF
